+++ core/dataSourceLocal/room/src/main/java/com/kychan/mlog/core/dataSourceLocal/room/DataSourceModule.kt
@@ -1,6 +1,7 @@
 package com.kychan.mlog.core.dataSourceLocal.room
 
 import com.kychan.mlog.core.dataSourceLocal.room.dao.MovieDao
+import com.kychan.mlog.core.dataSourceLocal.room.dao.MyMovieDao
 import com.kychan.mlog.core.dataSourceLocal.room.datasource.RoomDataSource
 import com.kychan.mlog.core.dataSourceLocal.room.datasource.RoomDataSourceImpl
 import dagger.Module
@@ -15,5 +16,5 @@ object DataSourceModule {
 
     @Provides
     @Singleton
-    fun providesRoomDataSource(movieDao: MovieDao): RoomDataSource = RoomDataSourceImpl(movieDao)
+    fun providesRoomDataSource(movieDao: MovieDao, myMovieDao: MyMovieDao): RoomDataSource = RoomDataSourceImpl(movieDao, myMovieDao)
 }
