+++ core/dataSourceLocal/room/src/main/java/com/kychan/mlog/core/dataSourceLocal/room/MlogDatabase.kt
@@ -5,6 +5,7 @@ import androidx.room.RoomDatabase
 import androidx.room.TypeConverters
 import com.kychan.mlog.core.dataSourceLocal.room.converter.SyncLogTypeConverter
 import com.kychan.mlog.core.dataSourceLocal.room.dao.MovieDao
+import com.kychan.mlog.core.dataSourceLocal.room.dao.MyMovieDao
 import com.kychan.mlog.core.dataSourceLocal.room.model.MlogMovieEntity
 import com.kychan.mlog.core.dataSourceLocal.room.model.NetflixMovieEntity
 import com.kychan.mlog.core.dataSourceLocal.room.model.SyncLogEntity
@@ -31,4 +32,5 @@ import com.kychan.mlog.core.dataSourceLocal.room.model.*
 )
 abstract class MlogDatabase : RoomDatabase() {
     abstract fun movieDao(): MovieDao
+    abstract fun myMovieDao(): MyMovieDao
 }
