+++ core/dataSourceLocal/room/src/main/java/com/kychan/mlog/core/dataSourceLocal/room/dao/MyMovieDao.kt
@@ -1,0 +1,29 @@
+package com.kychan.mlog.core.dataSourceLocal.room.dao
+
+import androidx.room.Dao
+import androidx.room.Query
+import com.kychan.mlog.core.dataSourceLocal.room.model.*
+import kotlinx.coroutines.flow.Flow
+import java.util.*
+
+@Dao
+abstract class MyMovieDao {
+    @Query("""
+            SELECT r.my_movie_id
+            , m.adult
+            , m.backdrop_path
+            , m.original_title
+            , m.poster_path
+            , m.title
+            , m.vote_average
+            , m.watch_providers
+            , m.rank
+            , r.rated
+            , r.comment
+            FROM my_movie AS m
+            INNER JOIN rated AS r ON m.id = r.my_movie_id
+        """
+    )
+    abstract fun getMyRatedMovies(): Flow<List<MyRatedMoviesVO>>
+
+}
